Securisation des actions sur utilisateur et articles

+++ src/controller/BackController.php
@@ -6,74 +6,97 @@ use projet_4\config\Parameter;
 class BackController extends Controller {
 
   public function addArticle(Parameter $post) {
+    if($this->checkAdmin()) {
 
-    if($post->get('submit')) {
-      $errors = $this->validation->validate($post, 'Article');
+      if($post->get('submit')) {
+        $errors = $this->validation->validate($post, 'Article');
 
-      if(!$errors) {
-        $this->articleDAO->addArticle($post, $this->session->get('id'));
-        $this->session->set('add_article', 'Nouvel article ajouté <br/>');
-        header('Location: ../public/index.php');
+        if(!$errors) {
+          $this->articleDAO->addArticle($post, $this->session->get('id'));
+          $this->session->set('add_article', 'Nouvel article ajouté <br/>');
+          header('Location: ../public/index.php');
+        }
+        return $this->view->render('add_article', ['post' => $post, 'errors' => $errors]);
       }
-      return $this->view->render('add_article', ['post' => $post, 'errors' => $errors]);
+      return $this->view->render('add_article');
     }
-    return $this->view->render('add_article');
   }
 
   public function editArticle(Parameter $post, $articleId) {
-    $article = $this->articleDAO->getArticle($articleId);
+    if($this->checkAdmin()) {
 
-    if($post->get('submit')) {
-      $errors = $this->validation->validate($post, 'Article');
+      $article = $this->articleDAO->getArticle($articleId);
 
-      if(!$errors) {
-        $this->articleDAO->editArticle($post, $articleId,$this->session->get('id'));
-        $this->session->set('edit_article', 'Article modifié <br/>');
-        header('Location: ../public/index.php?route=administration');
+      if($post->get('submit')) {
+        $errors = $this->validation->validate($post, 'Article');
+
+        if(!$errors) {
+          $this->articleDAO->editArticle($post, $articleId,$this->session->get('id'));
+          $this->session->set('edit_article', 'Article modifié <br/>');
+          header('Location: ../public/index.php?route=administration');
+        }
+        return $this->view->render('edit_article', ['post' => $post, 'errors' => $errors]);
       }
-      return $this->view->render('edit_article', ['post' => $post, 'errors' => $errors]);
+      $post->set('id' , $article->getId());
+      $post->set('title', $article->getTitle());
+      $post->set('content', $article->getContent());
+      $post->set('author', $article->getAuthor());
+      return $this->view->render('edit_article', ['post' => $post]);
     }
-    $post->set('id' , $article->getId());
-    $post->set('title', $article->getTitle());
-    $post->set('content', $article->getContent());
-    $post->set('author', $article->getAuthor());
-    return $this->view->render('edit_article', ['post' => $post]);
   }
 
   public function deleteArticle($articleId) {
-    $this->articleDAO->deleteArticle($articleId);
-    $this->session->set('delete_article', 'Article supprimé');
-    header('Location: ../public/index.php?route=administration');
+
+    if($this->checkAdmin()) {
+      $this->articleDAO->deleteArticle($articleId);
+      $this->session->set('delete_article', 'Article supprimé');
+      header('Location: ../public/index.php?route=administration');
+    }
   }
 
   public function deleteComment($commentId) {
-    $this->commentDAO->deleteComment($commentId);
-    $this->session->set('delete_comment', 'Commentaire supprimé');
-    header('Location: ../public/index.php?route=administration');
+
+    if($this->checkAdmin()){
+      $this->commentDAO->deleteComment($commentId);
+      $this->session->set('delete_comment', 'Commentaire supprimé');
+      header('Location: ../public/index.php?route=administration');
+    }
   }
 
   public function logout() {
-    $this->logoutOrDelete('logout');
-  }
+
+    if($this->checkLoggedIn()) {
+      $this->logoutOrDelete('logout');
+    }
+    }
 
   public function profile() {
-    $avatar = $this->avatarDAO->getAvatar($this->session->get('pseudo'));
-    return $this->view->render('profile', ['image' => $avatar]);
+
+    if($this->checkLoggedIn()) {
+      $avatar = $this->avatarDAO->getAvatar($this->session->get('pseudo'));
+      return $this->view->render('profile', ['image' => $avatar]);
+    }
   }
 
   public function updatePassword(Parameter $post) {
 
-    if($post->get('submit')) {
-      $this->userDAO->updatePassword($post, $this->session->get('pseudo'));
-      $this->session->set('update_password', 'Mot de passe modifié <br/>');
-      header('Location: ../public/index.php?route=profile');
+    if($this->checkLoggedIn()) {
+
+      if($post->get('submit')) {
+        $this->userDAO->updatePassword($post, $this->session->get('pseudo'));
+        $this->session->set('update_password', 'Mot de passe modifié <br/>');
+        header('Location: ../public/index.php?route=profile');
+      }
+      return $this->view->render('update_password');
     }
-    return $this->view->render('update_password');
-  }
+    }
 
   public function deleteAccount() {
-    $this->userDAO->deleteAccount($this->session->get('pseudo'));
-    $this->logoutOrDelete('delete_account');
+
+    if($this->checkLoggedIn()) {
+      $this->userDAO->deleteAccount($this->session->get('pseudo'));
+      $this->logoutOrDelete('delete_account');
+    }
   }
 
   public function logoutOrDelete($param) {
@@ -91,10 +114,14 @@ class BackController extends Controller {
   }
 
   public function administration() {
-    $articles= $this->articleDAO->getArticles();
-    $comments= $this->commentDAO->getFlagComments();
-    $users = $this->userDAO->getUsers();
-    return $this->view->render('administration', ['articles' => $articles, 'comments' => $comments, 'users' => $users]);
+
+    if($this->checkAdmin()) {
+
+      $articles= $this->articleDAO->getArticles();
+      $comments= $this->commentDAO->getFlagComments();
+      $users = $this->userDAO->getUsers();
+      return $this->view->render('administration', ['articles' => $articles, 'comments' => $comments, 'users' => $users]);
+    }
   }
 
   public function updateAvatar(Parameter $post) {
@@ -105,12 +132,36 @@ class BackController extends Controller {
   }
 
   public function unflagComment($commentId) {
-    $this->commentDAO->unflagComment($commentId);
-    header('Location: ../public/index.php?route=administration');
+
+    if($this->checkAdmin()) {
+      $this->commentDAO->unflagComment($commentId);
+      header('Location: ../public/index.php?route=administration');
+    }
   }
 
   public function deleteUser($userId) {
-    $this->userDAO->deleteUser($userId);
-    header('Location: ../public/index.php?route=administration');
+    if($this->checkAdmin()) {
+      $this->userDAO->deleteUser($userId);
+      header('Location: ../public/index.php?route=administration');
+    }
+  }
+
+  public function checkLoggedIn() {
+    if(!$this->session->get('pseudo')) {
+      header('Location: ../public/index.php?route=login');
+    }
+    else {
+      return true;
+    }
+  }
+
+  public function checkAdmin() {
+    $this->checkLoggedIn();
+    if(!($this->session->get('role') === 'admin')) {
+      header('Location: ../public/index.php?route=profile');
+    }
+    else {
+      return true;
+    }
   }
 }
